restruct cplusplus

+++ oj.ahstu.cc/1275.cpp
@@ -1,12 +1,9 @@
-#include <iostream>
-#include <iomanip>
-
+#include <bits/stdc++.h>
 using namespace std;
 
 int main(int argc, char const *argv[]) {
-  int a, b, c;
+  double a, b, c;
   cin >> a >> b >> c;
-  cout << setiosflags(ios::fixed) << setprecision(3) << (a + b + c) / 3.
-       << endl;
+  cout << setiosflags(ios::fixed) << setprecision(3) << (a + b + c) / 3 << endl;
   return 0;
 }
